mv Residual and SwiGLUFFN to block.py
